+++ opencvlib.h
@@ -1,4 +1,4 @@
 #ifndef OPENCVLIB_H
 #define OPENCVLIB_H
-void loadImage(unsigned char *&tex, int &sizex, int &sizey, const char file[], float times);
+void loadImage(unsigned char *&tex, int &sizex, int &sizey, const char file[], double times);
 #endif
